Add docker file to the repository
This took some inspiration by
[existing](https://hub.docker.com/search/?isAutomated=0&isOfficial=0&page=1&pullCount=0&q=stringer&starCount=0)
"Docker" builds.

But instead it uses "Alpine Linux" to provide a minimum
image. It also uses the multistage [build
feature](https://docs.docker.com/engine/userguide/eng-image/multistage-build/)
instroduced by the Moby team in 05/2017 to docker to ensure a minimal
image without any development dependencies (including the compiler to
build the gems). Since we need to run multiple commands on startup, it
uses foreman to monitor and kill properly on shutdown.

Before starting stringer, `bundle exec rake db:migrate` is run on each
startup. This is "bundle" in a small script. If one wants to run the
rake tasks themself, they need to overwrite the `CMD` of the container
on startup.